+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.6.12
 - Support Collection Exports
+- Disable UV editor in Level Design workspace (it causes crashes - in the future those may be fixed)
 
 # 1.6.11
 - Select vertex changes 'primary vertex' (enables merge at first and merge at last)